# lib/model, lib/osutil: Verify target directory before pulling / requesting