### Purpose

Ensure symlinks are not followed when pulling or requesting data.

This simply does an `lstat` walk of the components from the folder dir to the target directory, ensuring that all of them are real directories. This is of course not race safe, so assumes that the user does not locally replace a directory with a symlink at the same time as a pull is in progress against that directory. We should document this as a "don't do that". It is safe against ourselves, as we do not do directory/symlink replacement concurrently with pulling files.

This is cheaper than the stdlib `filepath.EvalSymlinks` which does a lot more work to find the actual absolute path of a file (which we don't care about).

It is also not cached at all. I don't think that matters in the puller, and I'm not currently measuring any difference in the request handling either, although I'm doing a couple of more tests. If it's an issue I think we can re-introduce fd caching in the request handler.

### Testing

A couple of new tests added for the pulling stage. It passes the existing integration tests for symlinks and basic syncing.