@calmh 2016-12-09
I added a benchmark for just a Request() call for a block in a file a few levels deep (it's in master).

```
master:
BenchmarkRequestInSingleFile-8   	  100000	     20330 ns/op	6447.12 MB/s

this branch:
BenchmarkRequestInSingleFile-8   	   30000	     42265 ns/op	3101.19 MB/s

with my secret fd caching:
BenchmarkRequestInSingleFile-8   	  100000	     10399 ns/op	12603.61 MB/s
```

Clearly, makes fast. But also clearly this is not the bottleneck, even at 42265 ns per call, given that we have a network roundtrip on top of that.