@AudriusButkevicius 2016-12-09
Where's the secret fd caching?