@Unrud 2016-12-10
This does not address the temporary file problem.
An attacker can create a Symlink like ``.syncthing.foo.tmp --> ../bar`` and then send an IndexUpdate for the file **foo** containing a manipulated version of **bar**. Syncthing will write to the file ``../bar`` outside of the synced folder.