@AudriusButkevicius 2016-12-10
There are no more temp files for symlinks.