@Unrud 2016-12-10
> There are no more temp files for symlinks.

Right, but you can still create symlinks that are named like temp files.