@AudriusButkevicius 2016-12-11
> I've tested it and was able to manipulate files outside of the folder.

Was this done on master with yesterday's commit?

> I've described one possibility but I may have missed something.

Yes the reordering would work, yet the iterator from the database always returns entries sorted lexically, so unless I am missing something I don't see how an attacker could force a different order coming from the db.

> How does it prevent the check if !osutil.IsDir(f.dir, filepath.Dir(file.Name)) { from being circumvented by ordering the children before the parents in the database on Windows?

I'd say because the graph performs ordering/sorting of it's own, irrespective of what order the remote device sends updates, yet given you are referring to windows specifically, I think I am misunderstanding something. Are you talking about someone sending a path with `\` in it from non-windows, to windows machine and then abusing that?