@Unrud 2016-12-11
> Was this done on master with yesterday's commit I've linked to?

5070d52f2f4763d363508d10c8f5f6eaf6e20646

> yet the iterator from the database always returns entries sorted lexically

Windows doesn't care for upper and lower case letters in paths. That can be used to sort children before parents like ``foo`` and ``FOO/bar`` because ``F < f``. I've not tested that, it's possible that I missed something.