@calmh 2016-12-11
> If the type of a Symlink on Windows is directory, IsDir() returns true. You have to check if the Symlink bit is set.

Fixed.

> Right, but you can create symlinks that are named like temp files.

Fixed. Also, the same check is performed before opening a file in the read-from-temporary path.

> That can be used to sort children before parents like foo and FOO/bar because F < f. 

Not fixed yet.

I am, again, starting to reconsider the value of supporting symlinks at all vs the pain they cause us...