@AudriusButkevicius 2016-12-11
That would be sad, given case insensitive indexes were on the list in the future anyway.