@calmh 2016-12-11
OK, how about this. While processing the Need list, we divide things into files (those get queued) and "everything else". The everything else pile gets sorted by number of paths components and then handled in that order. Once complete, files are handled.