@AudriusButkevicius 2016-12-11
Isn't that what we already do?