@AudriusButkevicius 2016-12-11
Ah the order is not by path component...