@calmh 2016-12-11
Yes, but there's not clear way around it to me. We already queue the files, this just adds the same for non-files.

One other way that doesn't work is to do this already in the database key format, that is instead or storing names as `foo/bar/baz` and sorting based on that we could have used `<2>foo/bar/baz` with an initial int16 or so for the number of path components. But that kills prefix scans and gets generally annoying.