@AudriusButkevicius 2016-12-11
Or we could use the spill sorter thing to create an intermediate index sorted by path component, yet I'd argue that it's perhaps better to make indexes non-casesensitive indexes as it squashes a few other bugs along the way.