@Unrud 2016-12-11
> I am, again, starting to reconsider the value of supporting symlinks at all vs the pain they cause us...

Just make a file system abstraction that is protected against symlinks (as I suggested previously).
All the security relevant code is one place and later you can even add platform specific code to protect against local symlink races.

> non-casesensitive indexes as it squashes a few other bugs

Make the file system abstraction case-sensitive on Windows.

> That would be sad, given case insensitive indexes were on the list in the future anyway.

No, don't submit to Windows.