@calmh 2016-12-11
Never start a sentence with "Just".