@calmh 2016-12-13
@st-review merge it for now. we will do more work in this area.