@st-review 2016-12-13
:ok_hand: Merged as 3582783972d37543d36b15c400bb89f3fd4f9bbb. Thanks, @calmh!